+++ pluperfect/lib/options.ts
@@ -23,46 +23,75 @@ export const DEFAULT_NO_CHANGE_COUNT: number = 0;
  * Results of parsing the command-line.
  */
 export interface CommandOptions {
-    /** true to force download of all files. */
+    /**
+     * true to force download of all files.
+     */
     force: boolean;
 
-    /** true if user requested help. */
+    /**
+     * true if user requested help.
+     */
     help: boolean;
 
-    /** true to enable JSON logging */
+    /**
+     * true to enable JSON logging. turns off regular logging.
+     */
     json: boolean;
 
-    /** spaces when rendering JSON. */
+    /**
+     * spaces when rendering JSON.
+     */
     jsonSpaces: string;
 
-    /** where to load standard locations. */
+    /**
+     * attempt to download live resources.
+     */
+    live: boolean;
+
+    /**
+     * where to load standard locations.
+     */
     locations?: string;
 
-    /** number of no change items processed before we stop. */
+    /**
+     * number of no change items processed before we stop.
+     */
     noChangeCount: string;
 
-    /** if true, only report errors. */
+    /**
+     * if true, only report errors.
+     */
     quiet: boolean;
 
-    /** flag indicating the message digest (hashes) should be recalculated. */
+    /**
+     * flag indicating the message digest (hashes) should be recalculated.
+     */
     rehash: boolean;
 
-    /** true if failures should be retried. */
+    /**
+     * true if failures should be retried.
+     */
     retry: boolean;
 
-    /** assume sync state (skip status checks when files don't change). */
+    /**
+     * assume sync state (skip status checks when files don't change).
+     */
     synced: boolean;
 
-    /** flag indicating more verbose output is desired. */
-    verbose: boolean;
-
-    /** flag indicating a request to print the version. */
+    /**
+     * flag indicating a request to print the version.
+     */
     version: boolean;
 
-    /** rest of the arguments of the command-line. */
+    /**
+     * rest of the arguments of the command-line.
+     */
     _: Array<string>;
 }
 
+/**
+ * @returns description of the command-line options to pass to the parser.
+ */
 export function getParseOptions(): ParseOptions {
     return {
         default: {
@@ -70,23 +99,23 @@ export function getParseOptions(): ParseOptions {
             help: false,
             json: false,
             jsonSpaces: '',
+            live: false,
             noChangeCount: `${DEFAULT_NO_CHANGE_COUNT}`,
             quiet: false,
             rehash: false,
             retry: false,
             synced: false,
-            verbose: false,
             version: false,
         },
         boolean: [
             'force',
             'help',
             'json',
+            'live',
             'quiet',
             'rehash',
             'synced',
             'retry',
-            'verbose',
             'version',
         ],
         string: [
@@ -116,10 +145,12 @@ export function printHelp(programName: string, parseOptions: ParseOptions): void
     console.log(`    output JSON structured log.`);
     console.log(`--jsonSpaces=spaces        [${parseOptions.default?.jsonSpaces}]`);
     console.log(`    spaces used to delimit generated JSON files.`);
+    console.log(`--live                     [${parseOptions.default?.live}]`);
+    console.log(`    download live resources.`);
     console.log(`--locations=name`);
     console.log(`    file to import for getStandardLocations() function.`);
     console.log(`--noChangeCount=number     [${parseOptions.default?.noChangeCount}]`);
-    console.log(`    number of items without changes before we stop.`);
+    console.log(`    number of items without changes before we stop the section.`);
     console.log(`--quiet                    [${parseOptions.default?.quiet}]`);
     console.log(`    be quiet. supress non-error messages.`);
     console.log(`--rehash                   [${parseOptions.default?.rehash}]`);
@@ -128,8 +159,6 @@ export function printHelp(programName: string, parseOptions: ParseOptions): void
     console.log(`    retry to download failed files.`);
     console.log(`--synced                   [${parseOptions.default?.synced}]`);
     console.log(`    assume repos start synced - limit checks for new files.`);
-    console.log(`--verbose                  [${parseOptions.default?.verbose}]`);
-    console.log(`    be verbose. include more informational messages.`);
     console.log(`--version`);
     console.log(`    print program version and exit.`);
 }
